pyproject.toml
[tool.poetry]
name = "varag"
version = "0.1.0"
description = "Create Vision Driven RAG Systems at Scale"
authors = ["Adithya S K <adithyaskolavi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "varag"}]  # Add this line

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
lancedb = "^0.13.0"
fastapi = "^0.115.0"
gradio = "^4.44.0"
openai = "^1.48.0"
transformers = "^4.45.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"